% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EdgeR.R
\name{build_contrasts_cluster}
\alias{build_contrasts_cluster}
\title{Functions for analaysing single cell data with EdgeR}
\usage{
build_contrasts_cluster(names)
}
\arguments{
\item{names}{Names of your clusters}
}
\description{
Functions for analaysing single cell data with EdgeR
}
