% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EdgeR.R
\name{build_contrasts_timepoints}
\alias{build_contrasts_timepoints}
\title{Build contrasts for timepoints}
\usage{
build_contrasts_timepoints(names, n_timepoints)
}
\arguments{
\item{n_timepoints}{Number of timepoints}
}
\description{
Build contrasts for timepoints
}
