% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EdgeR.R
\name{DEG_edgeR}
\alias{DEG_edgeR}
\title{Extract results from DEG}
\usage{
DEG_edgeR(
  qlf,
  p_value = 0.05,
  n = Inf,
  log_FC = 0.58,
  direction = "Up",
  genes_only = FALSE,
  all_genes = FALSE
)
}
\arguments{
\item{qlf}{Result from run_qlf}

\item{p_value}{P-value, default is 0.05}

\item{n}{Number of genes}

\item{log_FC}{Fold change, default is 0.58}

\item{direction}{Choose between up or downregulated genes}

\item{genes_only}{Returns only the gene names}
}
\description{
Extract results from DEG
}
