% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{dist_plot_highlight}
\alias{dist_plot_highlight}
\title{Plot cluster proportions with highlighted clusters}
\usage{
dist_plot_highlight(cluster, cols, percent = 50, data, split_by, fill_by)
}
\description{
Plot cluster proportions with highlighted clusters
}
