% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{is_outlier}
\alias{is_outlier}
\title{Functions to define outliers for QC}
\usage{
is_outlier(seurat_object, metric, nmads = 5)
}
\arguments{
\item{seurat_object}{Your seurat object}

\item{metric}{Mtric where you want to define ouliers eg mitochondrial genes}

\item{nmads}{Median Absolute Deviation}
}
\description{
Functions to define outliers for QC
}
