% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{marker_list}
\alias{marker_list}
\title{Extract marker genes as dataframe or gene list}
\usage{
marker_list(object, marker_genes, genes_only = FALSE, n = "inf")
}
\arguments{
\item{object}{Seurat object}

\item{genes_only}{Logical if True only gene names are returned}

\item{n}{Number of genes}
}
\description{
Extract marker genes as dataframe or gene list
}
