% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EdgeR.R
\name{plot_gsea_table}
\alias{plot_gsea_table}
\title{Plot GSEA table}
\usage{
plot_gsea_table(fgsea_res, msigdbr_list, ranked_genes)
}
\arguments{
\item{fgsea_res}{List with FGSEA results}
}
\description{
Plot GSEA table
}
