% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EdgeR.R
\name{ranked_genes_seurat}
\alias{ranked_genes_seurat}
\title{Rank genes by fc and p-value from seurat FindAllMarkers}
\usage{
ranked_genes_seurat(DEG_results)
}
\description{
Rank genes by fc and p-value from seurat FindAllMarkers
}
