% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EdgeR.R
\name{run_fgsea}
\alias{run_fgsea}
\title{Run FGSEA with a list of DEG results}
\usage{
run_fgsea(
  species = "mouse",
  category,
  subcategory,
  DEG_results,
  min_size = 15,
  max_size = 500
)
}
\description{
Run FGSEA with a list of DEG results
}
