% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{run_fgsea_edgeR}
\alias{run_fgsea_edgeR}
\title{Run FGSEA with a list of DEG results from EdgeR}
\usage{
run_fgsea_edgeR(
  species = "mouse",
  category,
  subcategory,
  DEG_results,
  min_size = 15,
  max_size = 500
)
}
\description{
Run FGSEA with a list of DEG results from EdgeR
}
