% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EdgeR.R
\name{run_fgsea_seurat}
\alias{run_fgsea_seurat}
\title{Run FGSEA with a list of DEG results from seurat}
\usage{
run_fgsea_seurat(
  species = "mouse",
  category,
  subcategory,
  DEG_results,
  min_size = 15,
  max_size = 500
)
}
\description{
Run FGSEA with a list of DEG results from seurat
}
