% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_object.R
\name{save_object}
\alias{save_object}
\title{Save your file without overwriting it}
\usage{
save_object(object_name, path_to_object, overwrite = FALSE)
}
\arguments{
\item{object_name}{Name of your object you want to save}

\item{path_to_object}{Path where you want to store your object, needs to contain the name for your object}

\item{overwrite}{TRUE for overwriting, default is FALSE}
}
\description{
Save your file without overwriting it
}
