% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EdgeR.R
\name{stats_edgeR}
\alias{stats_edgeR}
\title{Statistic for up and downregulated genes}
\usage{
stats_edgeR(qlf, p_value = 0.05, log_FC = 0.58)
}
\arguments{
\item{qlf}{Result from run_qlf}

\item{p_value}{P-value, default is 0.05}

\item{log_FC}{Fold change, default is 0.58}
}
\description{
Statistic for up and downregulated genes
}
