% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindMarker.R
\name{marker_list}
\alias{marker_list}
\title{Extract marker genes as dataframe or gene list}
\usage{
marker_list(
  object,
  marker_genes,
  genes_only = FALSE,
  n = "inf",
  per_cluster = TRUE
)
}
\arguments{
\item{object}{Seurat object}

\item{marker_genes}{Results of FindAllMarkers}

\item{genes_only}{Logical if True only gene names are returned}

\item{n}{Number of genes}

\item{per_cluster}{If TRUE marker_list contains df for each cluster}
}
\description{
Extract marker genes as dataframe or gene list
}
